feat: move lockfile logic to package utils
