+++ src/utils/package-json.ts
@@ -1,3 +1,5 @@
+import {existsSync} from 'node:fs';
+import {join} from 'node:path';
 import type {FileSystem} from '../file-system.js';
 import type {PackageJsonLike} from '../types.js';
 
@@ -19,4 +21,18 @@ export async function getPackageJson(
     // Not parseable
     return null;
   }
+}
+
+export const supportedLockfiles = [
+  'pnpm-lock.yaml',
+  'package-lock.json',
+  'yarn.lock',
+  'bun.lock'
+] as const;
+
+export function detectLockfile(workspacePath: string): string | undefined {
+  for (const c of supportedLockfiles) {
+    if (existsSync(join(workspacePath, c))) return c;
+  }
+  return undefined;
 }
